Add Codotto
I have added the platform that we have developed in conjunction with [ZgPHP](https://www.meetup.com/ZgPHP-meetup/) to help organize IT meetups. Codotto solely focus on IT community and it is completely free. Our features are only developed when we see that there is a direct benefit to the IT community of meetups.